Moved the claims component files.
