+++ token_manager/src/components/claims/ClaimView.tsx
