+++ token_manager/src/handlers/claims/TrustedClaimIssuerHandlers.ts
@@ -1,0 +1,20 @@
+import { ClaimType, TrustedClaimIssuer } from "@polymathnetwork/polymesh-sdk/types"
+import { IdentityGetter } from "../compliance/ComplianceHandlers"
+
+export type OnTrustedIssuerChanged = (trustedClaimIssuer: TrustedClaimIssuerFlat) => void
+export type OnTrustedIssuersChanged = (trustedClaimIssuers: TrustedClaimIssuerFlat[]) => void
+
+export interface TrustedClaimIssuerFlat {
+    identity: string
+    trustedFor: ClaimType[]
+}
+
+export const convertTrustedClaimIssuerToFlat = (trustedClaimIssuer: TrustedClaimIssuer): TrustedClaimIssuerFlat => ({
+    identity: trustedClaimIssuer.identity.did,
+    trustedFor: trustedClaimIssuer.trustedFor ?? [],
+})
+
+export const convertTrustedClaimIssuerFlatToCorrect = (getter: IdentityGetter) => async (flat: TrustedClaimIssuerFlat): Promise<TrustedClaimIssuer> => ({
+    identity: await getter(flat.identity),
+    trustedFor: flat.trustedFor,
+})
